requirements.txt
fastapi
uvicorn
python-decouple
SQLAlchemy
aiomysql
PyMySQL
PyJWT
passlib[bcrypt]
redis
python-multipart
alembic
